Remove tmpfile entries when removing file

+++ classes/task/remove_tmpfiles.php
@@ -46,11 +46,18 @@ class remove_tmpfiles extends \core\task\scheduled_task {
                   WHERE f.timecreated < ?",
                 array(time() - 86400)
         );
+        if (empty($tmpfiles)) {
+            return; // Nothing to clean up
+        }
 
         $fs = get_file_storage();
         foreach ($tmpfiles as $tmpfile) {
+            // Delete file
             $file = $fs->get_file_by_id($tmpfile->id);
             $file->delete();
+
+            // Remove tmpfile entry
+            $DB->delete_records('hvp_tmpfiles', array('id' => $tmpfile->id)) ;
         }
     }
 }
